type a and b multitouch is now supported
All of the event handling has been offloaded to the InputDevice threads
the poll_raw_input now only has to take some Vecs out of the channel, append them, and send them to the event loop.
